+++ CHANGELOG.md
@@ -6,6 +6,7 @@ BREAKING CHANGES:
 
 ENHANCEMENTS:
 * examples: add an advanced aws example to setup a Hopsworks cluster with ArrowFlight server
+* resource/hopsworksai_cluster: Add a new attribute `rondb/mysql_nodes/arrow_flight_with_duckdb` to enable or disable ArrowFlight server with DuckDB
 
 FEATURES:
 
@@ -21,6 +22,7 @@ ENHANCEMENTS:
 * test-fixtures: run against dev and prod envs
 * resource/hopsworksai_cluster: Set Default `version` to 3.3.0
 * Bump minimum Go version to 1.19
+
 
 FEATURES: